netCDF analysis Package
This is a comprehensive analysis package that enables users to find windturbine data closest to their desired coordinates and analyse the wind properties. A very breif example of a machine learning use to predict the power factor of wind turbine is also included.
A more comprehensive guide to the code will follow as soon as I get some spare time